+++ projects/Mallard/storybook/storyLoader.js
@@ -4,14 +4,14 @@
 // https://github.com/elderfo/react-native-storybook-loader.git
 
 function loadStories() {
-	require('../src/components/Lightbox/LightboxCaption.stories');
+	require('../src/components/Lightbox/Lightbox.stories');
 	require('../src/components/Spinner/Spinner.stories');
 	require('../src/components/SportScore/SportScore.stories');
 	require('../src/components/Stars/Stars.stories');
 }
 
 const stories = [
-	'../src/components/Lightbox/LightboxCaption.stories',
+	'../src/components/Lightbox/Lightbox.stories',
 	'../src/components/Spinner/Spinner.stories',
 	'../src/components/SportScore/SportScore.stories',
 	'../src/components/Stars/Stars.stories'
